Update README.md
Terro’s real-estate is an agency that estimates the pricing of houses in a certain locality. The pricing is concluded based on different features / factors of a property. The main objective of the project is to analyze the magnitude of each variable to which it can affect the price of a house in a particular locality . In my terros real estate project , i looked at different things like crime rates, distance and nox and other factors too figure out why property price changes . in this project i have applied some data analysis techniques such as reggresion modles , descriptive statistics and other statistical tools in excel to uncover the patterns and trends ummmmmm from this i dicovered interesting connections between the factors and housing cost. from this project i can conclude that neighborhoods with a higher percentage of lower income residents and lower number of room often correlated with lower propertyprices and houses with more rooms tendes to have higher average price .And from regression model we can say that crime rate is not a significant variable for average price of an house as p value is greater than 0.5. it is considered that p value less k than 0.5 is good and significant variable for performing regression models.and also nox , tax,ptratio and lstat have negative coefficient which says that increase in this features will result decrease in price of the house. The conclusion drawn from this analysis and This project helped me learn more about how these things work in real estate. I found that factors like tax , age and distance really impact house prices in the area. This project taught me a lot about real estate and data analysis using excel .